organizando funciones de post controller
